requirements.txt
panel
psycopg2-binary
sqlalchemy
pandas
python-dotenv